MAESTRO Phase 6: 47/67 GENESIS files compile (70%)
Additional fixes:
- EGG framework forwarding stubs (eggVector/Matrix/Quat → EGG/math.h)
- PlayerPointers resolved from game/kart/KartObjectProxy.hpp
- KartPhysicsEngine.hpp stub created
- System::Time and System::CharacterId enum guards
- BSP.hpp stub extended with BspWheel members (maxTravel, wheelRadius, contactDist)
- AudioEngine: alSourceStop stub, setCategoryVolume commented
- Collision: Field::ColInfo fully qualified, KartDynamics fromRotated fix
- Rendering: SceneGraph assignment fix, CourseObjAdvanced singleton
- Physics: PlayerSub10_status uintptr_t casts, field_0x196→field_0x190

+++ data/src/game/Rendering/CourseObj.cpp
@@ -176,7 +176,7 @@ void CourseObjActor_UpdateWithRotation(CourseObjActor* self, s32 variantIdx) {
     // Calculate pointer into the animation entry table
     // Entry offset = 0x2E0 + variantIdx * 0x0C
     u32 entryOffset = 0x2E0 + (variantIdx * 0x0C);
-    self->courseDataPtr = reinterpret_cast<u32>(static_cast<u8*>(courseData) + entryOffset);
+    self->courseDataPtr = reinterpret_cast<uintptr_t>(static_cast<u8*>(courseData) + entryOffset);
 
     // Set "loaded" flag (bit 14) on draw flags
     sub->drawFlags |= 0x4000;
